Fixes to posting problems in the graph-related globals
Modification to fzn_tree_int.mzn to prevent an inconsistency in the number of parameters received in the tree poster.
Modification of the steiner_tree poster so that the weight vector is initialized correctly.
Modification of the dag poster so that the new edges and node are added correctly in the relevant vectors. Also modified the nature of the new edges, as to not make every node mandatory. A bool_linear constraint is posted to deal with these new edges.