Correct an inheritance issue between the tree and the connected propagators.
ConnectedPropagator is defined as a subclass of TreePropagator in which some methods are redefined to deactivate the propagations relative to cycles. These methods were not made virtual so ConnectedPropagator had the same behavior as its parent.
Making the methods virtual lets the deactivation work. The new redefinition of the unite method is simply to remove an assertion regarding cycles that is probably useful when debugging the TreePropagator but becomes troublesome in ConnectedPropagator.