CMake: set version/soversion for libraries
For packaging, it is generally unsufficient to just have a `libXYZ.so`,
instead, that should be a symlink to the `libXYZ.so.SOVERSION`,
which, in turn, is a symlink to the actual `libXYZ.so.A.B.C` library.

I've looked at the previous pull requests, and it didn't look like
this was previously proposed, so i'm sending it...

Since building shared library is supported,
hopefully this is an acceptable change.
Currently effectively everyone has to carry this patch downstream.